requirements.txt
mcp>=1.0.0
notion-client>=2.0.0
aiosqlite>=0.19.0
pyyaml>=6.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.1.0
